+++ lib/libalpm/dload.c
@@ -247,7 +247,7 @@ static int download_internal(const char *url, const char *localpath,
 	while((nread = fetchIO_read(dlf, buffer, PM_DLBUF_LEN)) > 0) {
 		check_stop();
 		size_t nwritten = 0;
-		nwritten = fwrite(buffer, 1, nread, localf);
+		nwritten = fwrite(buffer, 1, (size_t)nread, localf);
 		if((nwritten != (size_t)nread) || ferror(localf)) {
 			pm_errno = PM_ERR_RETRIEVE;
 			_alpm_log(PM_LOG_ERROR, _("error writing to file '%s': %s\n"),
